moved old hamburger menu items to avatar display as an icon button and changed menu items for hamburger menu to match spec
